feat(ChangeRequest): add support for change requests to query repositories
In order go give users the capacity to either reject or allow a change based on shifts to a specified url, we need the concept of a "request" that can be either accepted or rejected. Accepting a request means adding the change to a requested history. Rejecting it should mark the change rejected in some way.