chore(dependencies): vendor in all dependencies
So this has been quite a nightmare. I ended up using `govendor add`
to manually ingest all dependencies by repeatedly building in our
test environment. The upside is a passing set of tests and a much
clearer build path. The downside is this is going to make changing
qri subpackages tougher (possibly a good separation of concerns),
and making dependency management much tougher
to manually ingest all dependencies by repeatedly building in our
test environment. The upside is a passing set of tests and a much
clearer build path. The downside is this is going to make changing
qri subpackages tougher (possibly a good separation of concerns),
and making dependency management much tougher.
I think the proper long term solution is to have a good look at gx
and see if we can't submit some improvements that increase
compatibility.