test(core.DatasetRequests.List): add tests to check sort order of DatasetRequest.List response
lot's will need to land before we have a proper fix to the ordering of datasets, at least this
test makes it possible to check when the api is behaving as expected. In the future this
test should be moved down into the repo.Namespace() method, where this logic should actually
be happening.