refactor(core.DatasetRequests.Update): refactored and improved update method params
a few enhancements and breaking changes to the update method. We now (finally) accept
a Reader of data for updates (lots to do here, including schema validation). Now that
commits are properly integrated with datasets, update accepts a candidate dataset that
contains all changes to the dataset, including setting the Commit and Previous fields
directly on the updated dataset itself.
made use of new dsfs.RefType method to open up the number of ways one can refer
to "Previous"