feat(RPC): change default port, provide RPC listener
so we've been planning to do this for a while, but a proper
reason to do so hadn't come up until we needed to be able to
use the CLI while a server is running (while ssh'd into a container).

This commit is a spike that only supports DatasetRequests for now
while I work out the details of this pattern.

also, closes #163