feat(api.ServeRPC): serve core methods over RPC
this commit builds upon intial work to make all core methods available
as RPC. this isn't finished as of yet, given that all in and out params
will need to conform to encoding/gob specs, which basically boils down
to using only Exported types. I think this is a good opportunity to
have *all* params to core methods be declared in the core package
itself (these could be aliases to underlying types if needed).

TODO:
 * finish implementation
 * write test suite that works core methods over RPC