test(p2p): built initial message test
This started out as trying to debug a connection error that turned
out to be the result of leftover config.json files from the switch
to config.yaml being read in instead, but now we have a PING message,
and a test that spins up a swarm of at least 2 qri nodes, testing
actual message passing between them.

From here on in I think all enhancements to our p2p system should
start with a test.