feat(license): switch project license to GPLv3
This project is currently small, only known to few. But even so, the
group of people who have helped get qri to this point deserve to know
that no matter what the future holds, they will have a right to use
what we're making here.

I'm leaving all of qri's subpackages as MIT in hope that if there's an
off-label use for anything we're building, others should be able to
pursue that use, and make their own licensing decisions.

When it comes to our cornerstone project, however, I agree with Liz.
I think we're building has a chance of becoming public infrastructure,
and I want make sure it stays there.