refactor(repo, core): DatasetRefs errwhere, Namestore -> Refstore
Our version of dataset names that we've used so far landed pretty late
in the course of shipping 0.1.0. We've set the external expectation
that peername/datasetname is a proper way to refer to a dataset, but
haven't enforced that expectation across the codebase.
The most obvious place this shows up is the repo.Namestore interface,
which makes less sense in a world where we use repo.DatasetRef to move
dataset names around.
With that in mind, I've renamed Namestore to Refstore and changed it's
interface to reflect this. The most notable change is GetRef, which
should probably be renamed to "CompleteRef" in the future, as that's a
more accurate description of what it does. If given a peername/name combo,
GetRef adds the path, if given a path, GetRef adds the peername/name.
This removes a *lot* of code before that used to be `if ref.Path == ""`
before calls to dsfs.LoadDataset.
In the future we can consider making GetRef an internal method to
be used by higher-order repo interface methods. For example
repo.GetDataset would internally call repo.GetRef, allowing callers of
repo.GetDataset to just pass in incomplete references, expecting the repo
to work it out for them. This'll de-duplicate a bunch of code in core.