feat: reference canonicalization
While giving users the capacity to say "me" instead of "peername" is nice, it creates a number of dangers. We need to make sure that no dataset is ever stored as `me/dataset`, as that'll collide like crazy. I'm calling this process of `me -> [peername]` _canonicalization_, because it's getting a dataset name into it's canonical form. We need to do this process as early as possible. Doing this canonicalization, however, requires knowing what the current peername is, and that's stored in a repo. So I've added a new method to the Repo interface: `ParseDatasetRef`. Implementations of repo are expected to use another new method: `repo.CanonicalizeDatasetRef` to turn any dataset reference that uses "me" (or the edge case of a blank peername) to the canonical name for the reference. While we're in there, `CanoncializeDatasetRef` also tries to proactively set the reference path using any info in the repo. This gives us a great starting point for checking when we do and don't need to connect to the network.

However, This repo requirement creates a number of headaches for RPC, so I think the next commit will refactor _when_ we canonicalize references to post-rpc points in core methods, when we're garunteed to have a repo instance.