refactor(save, api/init): refactored core.Save, cmd/save.go to match patterns found in core.Init, cmd/add.go
api.initHandler and api.saveHandler needed to be refactored to only accept files via mime/multipart

ran into some interesting issues while refactoring the core.Save function: \n-currently, if a user adds metadata or structure, but does not add data file, we load the data from the previous commit. this means that the SAME data is getting saved to the store, which is bad\n-when we use the assign function, we assign incorrect paths to the metadata and structure strucs, the current metadata and structure incorrectly have the same paths as the previous metadata and structure. when we compare these current and previous structs with datasetDiffer, the differ returns no diffs, since it compares the paths and finds them equal. because of this, if we dont add data and dont add a commit title or message, then the commit will have no title or message, since the differ will not find any differences between the two versions