fix(save): reset the meta and structure paths after assign
now that we have the handy new SetPath function, we can reset the meta and structure paths after assign. assign clobbers the changed paths with the previous paths. this means that the datasetDiffer will assume no changes have been made in the meta or structure. we want the datasetDiffer to have to check each field, so we reset the path