refactor(profile): profile has separate function for saving peername
In the future, when we move away from peername as a source of truth and toward ID, we can probably get rid of this. For now, however, let's make sure that, for safety, the only way to save a peername is via the SavePeername function

We also now have a ValidatePeername func that makes sure that the proposed peername is not on the reserved names list. The skeleton for ValidateProfile has also been laid.