feat(JSON): lots of JSON support fixes. updated to go 1.10
this proves qri can do JSON. We've also added the capacity to have
the top level of a dataset be an object.

unfortunately, it looks like I've shipped a small regression in
dataset schema detection, which seems to now point out booleans
as strings. I'll file an issue and fix.