feat(webapp): fetch webapp hash via dnslink
So, we want to be able to modify & update the webapp independent of qri,
to do that, I've created a new dnslink entry "webapp.qri.io" that can be
resolved with `ipns name resolve webapp.qri.io`. For kicks I've also added
ipns name resolving to the api server (this should be removed when we refactor
to just using regular IPFS API bindings).