fix(CanonicalizeDatatsetRef): should hydrate datasetRef if given path
If we gave it a datasetRef with only a path, CanonicalizeDatasetRef
would not try to find the peername, peerid, or datasetname associated
with that path. This isn’t perfect, as it only checks the refstore for
a listing of that path. If it finds a listing, it checks to make sure
that any info we have been given matches the info found.
If no ref is found, we just return as is.