fix(Validate): completely overhaul validation to make it work properly
So, turns out a *bunch* of stuff needed fixing when it comes to valiation.
We've gone through and cleanup up the following to make validation a better experience.

jsonpointer:
* added a new "Descendant" method on a pointer to generate child paths

jsonschema:
* complete overhaul of Validation interface to support multiple errors, consistent error
  formatting, and added a jsonpointer path to the location of each error

dataset/dsio:
* added a new "Entry" struct to represent top-level entries in a dataset
* overhauled dsio to remove vals package. dsio now works with go prinitive types
* CSVReader has been completely reworked to convert entry rows to the types specified
  by the Structure schema of the reader (this was the root cause of #290)

dataset/validate:
* added a new Data method that consumes a dsio.EntryReader to validate data that auto-converts
  incoming data to JSON for validation (a necessary step to work with JSONSchemas)

dataset/dsfs:
* dsfs.CreateDataset now uses validate.Data

dataset:
* updated all packages to use new dsio interfaces

qri/core:
* dataset.Validate now is mainly a wrapper around validate.Data

This sets the stage for some exciting stuff in the future, and makes qri work properly / better today.
I'm particularly excited to work with the new dsio Reader/Writer interfaces, as they set the stage for
filtering & sorting subsections of datasets on the fly

closes #290