feat(repo.Event): removed DatasetAnnounce in favor of event logs
I've quickly come to realize that the DatasetAnnounce approach leaves much to be
desired. Questions about when to announce, and ensuring that both peers of a shared
profile catch the message is just too much. Instead I've refactord out the QueryLog
interface from Repo, and replaced it with a log of "events" which record public actions
a peer has taken. This will make sync a much easier affair, where two peers with a matching
profile will discover each other, do a security handshake, and then sync their events
based on peer settings.

It seems there's a decent amount of pressure mounting to do a heavy refactor of the Repo interface,
pushing parallels of core's business logic methods into the Repo interface (repo.CreateDataset is the
leading-edge example).