feat(actions): added new repo actions package to encapsulate repo biz logic
For some time I've been longing for a package that encapsulates the business logic for "doing stuff
with repos". This is lower level than core (which has to deal with the question of when to
touch the p2p network), and higher level than repo implementations to avoid duplicating this
code everywhere.

This actions package comes from refactoring methods like CreateDataset _out_ of the Repo interface,
and adjusting these business logic-heavy methods to work only with Repo interface methods. Already
feels better.

I've named the package actions for anyone who's coming to qri from the world of react & redux,
where the concept of an action plays a similar role. In this version actions don't emit events
to work on stores, but instead work directly on stores (repos) to do their work.

This analogy gives more insight into the role repos should play moving forward. In frontend dev,
the store is _a_ store of truth, but the role of being _the_ store of truth is always the server.
In our case, repos serve a similar purpose, only this time the "server" is the distributed web.

In both cases the store/repo is comparitively cheap to query, at the expense of being potentially
out of date. So it kinda like a "necessary cache".

Dope.