refactor(profile.ID) adding peer / profile ID distinctions
for a while now we've been conflating two concepts that need to be properly pulled apart. We could
give these concepts a number of names, so far the best I can come up with is profile and peer.

In this context, a *profile* is the information about a given person or group.

A *peer* is a running qri program connected to other qri programs.

A profile can have many peers. Peers can only have one profile (for now at least). The purpose of
this distinction is to seperate social identities (profile) from network identities (peers). This is
to deal with the reality that people have multiple devices.

It's worth noting that I plan to still refer to the combination of a peer & profile as a peer.
This is confusing to those who need to know the distinction but important for social purposes.