test(cmd,api): skip integration tests if qri is already running
running tests when a qri node is already running can cause bad things to
happen, so I've added checks & skips if a server is already running.

I've chosen to skip instead of fail b/c it's nice to still be able
to run other tests while a qri node is up, but this can (and probably will,
lead to false positives), thankfully CI should catch this stuff, as it runs
in a vacuum.

closes #342