feat(cmd.Config): added qri config export command
while testing out enviorment variable configuration I needed a quick way to get config an existing
config as JSON, and realized it may be helpful in other places. This adds export as a subcommand of
config that by default prints a json blob of the current config file with private keys removed. I've
added --with-private-keys to add in the private keys, and a few other flags to switch format to yaml
or pick an output file.