refactor(config): move global config into core, refactor CLI config loading
Thanks to storing profile details in config, we now need a clean, canonical
method for manipulating configuration files to update profile details.

The first step to dealing with this is getting config, and the path to
where the config file is stored in core, so that both API and cmd packages
can manipulate the right file.

While we're at it, I've changed loading configuration to be a PreRun hook
on commands that need access to configuration (which is many). This makes for
easier integration with cobra by placing explicit loadConfig calls, only when
needed.

This'll also be put to use when we ship frontend features for configuring your node.