fix(connect): fix errors caused by running connect with setup
Version 0.3.0 went out the door with a bug that breaks the default
docker image command, sad! The problem was caused by switching the
"anonymous" flag from a cobra.BoolVarP to a cobra.BoolP.

When connect is run with --setup, we cheat heavily & just manually
invoke the setup cobra command on the spot. The problem with this
cheat is the setup command hasn't been properly allocated by cobra,
and so any flag set without a global variable reference isn't assigned
when commands are invoked in this manual way. In the long run we should
move away from this method entirely. The short solution would be to move
the contents of setupCmd.Run into a private func "doSetup", the better
way to do this would be to create a new setup core method.

To smooth the transition through these refactorings, I've added a test
for the connect command that confirms it's behaving.

The only problem with testing the connect command is connect is designed
to be a long-running process.

 So to get around this I've added a new configuration option to
API config, and argument to connect for "disconnect after", which
tells the server to shutdown after a designated number of seconds.

We might be able to use this in other places, I'm thinking of cron
jobs that spin up with a very long disconnect-after to ensure they
don't run forever, terminiting before the disconnect-after deadline
when their work is confirmed finished.