feat(registry): use registry at setup to claim peername
Now we can finally avoid peers accidentally claiming the same peername.
To get this to work I've moved setup into the core package. Setup is going
to get a little more sophisticated in the future, and I'd prefer to have
the core method in a place where it can be tested independent of the CLI,
even if the only way to access this codepath from userland is through the
CLI.