feat(teardown): added 'uninstall' option from the CLI & core Teardown method
Now that registries are added, we need a clean way to tell the server we want to
opt out of using qri. We can now do that with `qri setup --remove`.

While in here I've removed the manual setup command invocation from the connect
command in favour of a doSetup func. The manual invocation was a source of bugs &
pain, so I'm happy to get rid of it even if what's replacing it could use more work

closes #175