refator(api): use /profile/photo & /profile/poster as vanity enapoints
Similar to the situation we ran into shipping the webapp, readonly mode
can and should prevent access to anything under the /ipfs/ endpoint,
which sadly includes profile photos & poster images. To get around this
we'll use the same trick of hiding photos behind vanity urls.

In an effort to avoid API bloat, I've added these as elements of
the previously under-used /profile/photo & /profile/poster GET endpoints.
Both now accept query params for id (which is the profile ID) or peerneme.
this is going to need some refinement & docs, but it'll work for now.