fix(core.Add): major cleanup to make add work properly
now that we have a clean story around CodingDataset, it's bene much easier
to clean up the add command & get it working over RPC. There were/are a *bunch*
of lingering issues, as adding a remote dataset touches every major feature of
our infrastructure.

While fixing add, the issue of "who owns a dataset" has come up in a big way, as
two lists are now in contention: the list of dataset's a profile has *created*,
and the list of datasets a peer is *storing*, which can be different. As an example,
if my profile is `b5` and I run `qri add edgi_webmon/epaGovSitemap`, I'm saying
"give me that dataset made by edgi_webmon". Based on that, running `qri list` with
no arguments now returns the _full list of stored references_, which in this case
will include `edgi_webmon/epaGovSitemap` in the list of output datasets. To see the
datasets I have created, I now have to run `qri list me` or `qri list b5`. I'm not
totally sure how I feel about this, but I think this is a usable solution for now.

All of this needs more thought & tests. In the near future we should definitely look
to add tests that simluate at least two qri nodes running core methods that require
interaction with each other. I'm thinking this'll make for a nice core package
integration test.

The other thing I'd like to see is adding datasets & histories to registries,
sewing in a fallback during add that checks a registry during DsRef canonicalization.
As histories accumulate we're going to see bugs and confiusion resolving which version
is tip, as the references themselves contain no mechanism for determining tip without
inspecting histories.

There's so, so much work to do here, but I'm very happy to at least see this thing
working