feat(p2p): Tag qri peers in connManager
When operating on "real" IPFS networks, we need to let the connection manager know
that peers supporting the qri protocol have more value to us. I've added a default
value of 6 as a starting point. If anyone's wondering how hard I rep Toronto,
HERE'S YER PROOF.

One day when we have time to deal with stuff like this I'd love a test that checks
for priortization of connetions amongst a number of qri & non-qri peers. There should
be situations where we choose to deprioritize a qri peer. One thing we may do is
use reputation coming from registries we've opted into to tag bad-actor peers with a
*negative* tag value.

Another thing that would be nice is to go out & collate tag values others are using
I'm currently seeing "kbucket":5, and "relay-hop":2.