feat(registry.Datasets): initial dataset registry integration
we gave registries a few new tricks, and have started integrating them here.
If a qri repo is configured with a registry, it will fire off notifications
to the registry server within goroutines on dataset creation & deletion, which
will start to build up a list of datasets on the registry side. This comes back
to help users on core.datasetRequests.Get, which now includes a parallel check
to both p2p network *and* registries if we don't have a local copy.

The upside here is we can use registries to do checks if we aren't connect to p2p.
If this goes as expected the next place to expand is adding search on the registry
side, and then move on to tracking who has pinned what.

*might* make some inroads on #397