feat(add): accept --dataset yaml/json file arg to add
to make the addition process easier, I'm adding a --dataset arg to the add command that takes a  "DatasetPod" object, which allows users to specify the pieces of the dataset they'd like to provide in one file. More documentation on this to come.

To do this the "right way" I've put a bunch of refactoring into the core "init" method, making it possible to use `add` while the api is running (by passing local filepaths instead of io.Readers in input params). This overhaul is faster, bumps test coverage up, and consumes less memory (even with a fancy io.TeeReader!).