refactor(core.Add, core.Save): refactor input params, folding into a DatasetPod
This change is a first example of eliminating the need for DatasetRef by leveraging fields present in dataset.DatasetPod. DatasetPod contains all the fields necessary to submit changes to a dataset, so Add & Save methods can now be refactored to accept a file that fully encapsulates the dataset to create.

BREAKING CHANGES:
add & save on both API & CLI now accept a "file" which is the full dataset