feat(transform secrets): supply secrets to transforms
At some point someone isn't going to read the documentation on how transform.config works, not realize that
we record all configuration details, create a script that asks a user to put API keys into a config slot,
and qri will dox those keys. This will make me sad.

To prevent this I'm adding in an explicit codepath for secrets, one that currently only works with the
command line, and comes with a warning if the user tries to provide secrets to `qri add` or `qri save`.
By providing explicit features & docs on how to interact with secret information, and throwing warnings
around the use of this information, I believe we'll reduce the chances people will misuse features like
config. @osterbit pointed to the "danger zone" on github repo settings as a good example of warning users
when they're about to do something potentially dangerous.

Secrets will never be recorded, and will need to be provided on each run of a transform. Later on we can
think about some sort of secure store of secrets that can be auto-provided to scripts, which we can then
plumb up into UI.

Because users are required to interactively confirm when providing secrets, I've also added a `--no-prompt`
flag to skip any prompts. Advanced users will need this for scripting purposes. I've intentionally left off
a shortcut so that users must explicitly type "no-prompt".