refactor(rename core): rename core to lib
I'm now starting to use qri core outside of qri for our tutorials project. I want to start thinking about using qri in this context, how we should be conceiving of core as an importable package, what interfaces it should present, etc. Best place to start is with renaming core to lib, which a few of us have been discussing for a while now.