feat(lib.Error): new type Error that satisfies the error interface
lib.Error wraps and `err` error and a `msg` string.  lib.Error gives us an opportunity to couple dev focused error messages with more user-friendly errors. Before returning an error message to the user, we can check if it actually a lib.Error. If it is, we can send Error.msg to the user instead.