test(cmd/render, cmd/test_factory): fill out tests for Run func in render command
In order to test that render works, we need to supply a template.

There are two ways a user can supply a template, through the default template, whose address lives in the config file, or by passing a filename

For the first, we had to construct a MemFile, add that file to the store, and update the config with the key to that file. When the lib.Render function sees that no template is given, and if the lib.Config struct is set, it will grab the key to the template (Config.Render.DefaultTemplateHash), and use that to get the template from the repo.Store.

For other testing purposes, we have a new template.html file in the cmd/testdata/ folder, which has a few basic template params.