test(server): fix searchResponse.json
So there was a point when the search response actually responsed with the correct search value. Now it's returning with an empty data field.

Definitely need to figure out why, but I've been spending too much time on this and I want to progress, so for now keeping the weirdness.