refactor(dataset): Change CanonicalizeDatasetRef's return semantics.
Previously, CanonicalizeDatasetRef had 2 main steps: canonicalizing the profile
(turning "me" into the user's peername), and then filling in fields from the
local repo. If the dataset did not exist in the local repo (because it was
a remote dataset), the function would ignore GetRef's error, and just continue
on. This fragile behavior led to bad things, like many callers duplicating work
such as caling GetRef after CanonicalizeDatasetRef, or certain methods
displaying non-sensical error messages.

Instead, have CanonicalizeDatasetRef return repo.ErrNotFound if the reference
is non-local. Callers can handle this error code as needed, such as displaying
a more sensible error, or succeeding anyway (in the case of `rename`, which
expects the reference to not be found), or contacting a remote host (like
`log` does to get dataset about non-local references).