refactor(actions): remove actions structs in favour of Exported functions
the actions package initially oriented around structs, where consumers would allocate,
for example, an `actions.Dataset` struct that would embed a reference to repo.Repo,
allowing actions to mix repo.Repo method calls with action struct methods. In
practice I found this confusing & irritating.

This refactor clarifies the purpose of the actions package. actions are a collection
of canonical business logic functions, mainly meant to be consumed by the lib package.
Orienting around functions makes for longer method signatures, but allows each function
to be specific in what it needs and clear that it's not relying on any internal state or
structures to execute. I expect this package to grow, and the comments on each function
to grow with it.

I've also added a bare minimum of tests to make sure each exported test is being
worked at least once. There's lots to do before this test suite is anywhere near
complete, but at least it's not embarassing anymore. The test suite as it is could
do for a round of refactoring for deduplication.