fix(peer sharing): fix peers not sharing info with each other
Trying to make inroads on #510, I've re-worked the way peers share info about
each other. Instead of asking for profile information, a new request method "qri_peers"
asks for just the list of peer & profile id's a connected peer is connected to.
From there the local peer examines the list, and attempts to connect to any addresses it
hasn't seen before. This process repeats until no new peers are found.

Lots of work left to do here, but this should be a massive improvement