refactor(QriNode): add registryConfig to the QriNode
To be able to call to the registry to get the reputation of a qri peer, we need a registry location. So instead of passing in just the `*config.P2P` we pass in the whole `*config.Config`

Also, in `makeBasicHost`, we add a ConnectionManager to opts. If we don't do this, our `BasicHost` has a `NullConnManager`, which does nothing.