tests(p2p): adjust tests to not rely on AddQriNode
So previously, AddQriNode was erroneously adding Addrs, and tagging.

AddQriNode should not have anything to do with the act of connecting two nodes, it should just on board a node we have already connected to as a qri node.

In order to connect two nodes, we need to call `n.Host.Connect(context, peerinfo)` and then handle the connection: `n.HandlePeerFound(peerinfo)`

This should probably be put into a convenience test function.

Also, we should probably refactor `n.ConnectToPeer(ctx, peerParams)` to take peerinfo, and do all the stuff that would happen if we found a peer in the wild (aka, `HandlePeerFound`. My mental model may be off, but it seems that if we `n.Host.Connect()` to a peer, we don't actually run through the `HandlePeerFound` function (which tags the peer as a Qri peer).

Anyways, refactored the tests so that it was connecting to peers in the correct (as I see it now) way.