refactor(p2p): rename `ConnectQriPeers` to `ConnectQriNodes` & adjust tests
this rename mirrors our `ConnectNodes` function. `ConnectNodes` creates a basic connection between each node. `ConnectQriNodes` ensures that this connection is a Qri connection.

Adjust tests to use `ConnectQriNodes`

The only thing that is sort of worrisome is that `ConnectNodes` is probably also triggering `UpgradeToQriConnection`, when `HandlePeerFound` is called. So we may never actually be in a state where `ConnectNodes` is actually giving us back basic connections.  Just a thought.