refactor(api.saveHandler, lib.Save, base.CreateDataset): add ability to use DryRun and ReturnBody flags
We haven't yet moved the initDataset functionality into SaveDataset.

In the API, we ignore the `dry_run` and `return_body` form fields, so all calls to save from the API are actually saving the dataset, not actually doing a dry_run

This code does a hacky work around on the MemRepo and loading a previous dataset known bug.

If we ReturnBody is true, we return the dataset body as as bodyBytes, using the addBodyFile function