refactor(Save): complete overhaul of save path
I'm tired now. This was hard, but well worth it in the end. I'll follow up with shoring up some tests, and I've found some bugs in other packages during this process, but here's the general guide I was using, which was a refinement on what we worked in RFC0017, refined based on what we learned in RFC0020:

lib.DatasetRequests.Save                Prepare
  actions.Recall                        |
  actions.SaveDataset                   |
    base.PrepareSave                    |
      prev = base.LoadPrevious          |
      prev.Commit = nil                 |
      prev.Transform = nil              |
    prev.Apply(ds)                      Transform
    if ds.Transform                     |
      ExecTransform                     |
    base.CreateDataset                  |
      InferValues                       Compute Automatic Values
      SetComputedValues                 |
      Validate                          Validate
      dsfs.CreateDataset                Create