refactor(p2p.ClosestConnectedPeer): rename and refactor to `ClosestConnectedQriPeer`
our ClosestConnectedPeer function attempts to:
1) get the peerIDs of the profileID in question. If we don't have these ids, or if we don't have connections to these ID we move on to step 2
2) we iterate through our connections and create a slice of the peerIDs we do have connections to, up until a given max number of connections.

However, since we are trying to contact these peers using the Qri protocol, we should actually check to see if these peers speak qri before adding them to the list of peers we want to contact.

In the future, we will probably want to come up with a better way to determine which qri peers are actually the closest to the qri peer in question.