refactor(registry): replace Status with RegistryDataset
We want to use the `/registry/` GET endpoint for RegistryDataset, which will essentially do the same thing as status, except it will return the dataset instead of a bool.

Made a change to `repo/ref` `CanonicalizeProfile`, taking advantage of the `ErrNotFound` error. In certain cases we want to forge ahead even if the dataset or profile has not been found locally.