docs(*): updating errors in docs
Looks like we never had a proper CONTRIBUTOR.md file like we do in the frontend. This modifies that file for use in the backend. Also, our DEVELOPERS.md file seems like it was just copy pasted over to the backend github, without changes that make it relevant for go

closes #619