fix(cmd): Both body and setup runnable during "qri connect"
Prior to this change, both `qri body` and `qri setup` would fail with
"repo not available (are you running qri in another terminal?)" if run while
`qri connect` was running. This is due to the face that the repo and ipfs are
locked so that only only process touches that at once. The Repo() method on
cmd/factory.go Factory would therefore fail with this error. However, body
doesn't actually need to directly touch the repo, and can be run over RPC
like other commands. Setup will fail anyway because the repo already exists.

Only `qri export` is now using the Repo() method, we should stop doing that
and then remove Repo() if possible.